feat: Enhance calendar event management by adding event creator functionality - added new event_creator column to meeting table

+++ apps/meet-app/backend/service.bal
@@ -373,11 +373,11 @@ service http:InterceptableService / on new http:Listener(9090) {
                 body: {
                     message: customError
                 }
-            }; 
+            };
         }
         // Attempt to create the meeting.
         calendar:CreateCalendarEventResponse|error calendarCreateEventResponse = calendar:createCalendarEvent(
-                createCalendarEventRequest, userInfo.email,meetResponse);
+                createCalendarEventRequest, userInfo.email, meetResponse);
         if calendarCreateEventResponse is error {
             string customError = string `Error occurred while creating the calendar event!`;
             log:printError(customError, calendarCreateEventResponse);
@@ -393,7 +393,7 @@ service http:InterceptableService / on new http:Listener(9090) {
         string rule = "";
         int[] meetingIds = [];
         if isRecurring {
-            gcalendar:Event|error masterEventResp = calendar:getCalendarEvent(calendarCreateEventResponse.id);
+            gcalendar:Event|error masterEventResp = calendar:getCalendarEvent(calendarCreateEventResponse.id, userInfo.email);
             if masterEventResp is error {
                 string customError = string `Error occurred while getting master event!`;
                 log:printError(customError, masterEventResp);
@@ -409,7 +409,7 @@ service http:InterceptableService / on new http:Listener(9090) {
                 rule = recurrence[0];
             }
 
-            gcalendar:Event[]|error instances = calendar:getEventInstances(calendarCreateEventResponse.id);
+            gcalendar:Event[]|error instances = calendar:getEventInstances(calendarCreateEventResponse.id, userInfo.email);
             if instances is error {
                 string customError = string `Error occurred while fetching recurring instances!`;
                 log:printError(customError, instances);
@@ -477,6 +477,7 @@ service http:InterceptableService / on new http:Listener(9090) {
                     title: originalTitle,
                     googleEventId: instance.id,
                     host: userInfo.email,
+                    eventCreator: userInfo.email,
                     internalParticipants: string:'join(", ", ...createCalendarEventRequest.internalParticipants
                             .map(internalParticipant => internalParticipant.trim())),
                     startTime: startTimeDb,
@@ -513,6 +514,7 @@ service http:InterceptableService / on new http:Listener(9090) {
                 title: originalTitle,
                 googleEventId: calendarCreateEventResponse.id,
                 host: userInfo.email,
+                eventCreator: userInfo.email,
                 internalParticipants: string:'join(", ", ...createCalendarEventRequest.internalParticipants
                         .map(internalParticipant => internalParticipant.trim())),
                 startTime: createCalendarEventRequest.startTime
@@ -658,9 +660,21 @@ service http:InterceptableService / on new http:Listener(9090) {
             };
         }
 
+        string|error? creatorEmailResult = database:getEventCreatorByGoogleEventId(meeting.googleEventId);
+        if creatorEmailResult is error {
+            string customError = "Error occurred while fetching the event creator!";
+            log:printError(customError, creatorEmailResult);
+            return <http:InternalServerError>{body: customError};
+        }
+        if creatorEmailResult is null {
+            string customError = "Event creator not found!";
+            log:printError(customError);
+            return <http:InternalServerError>{body: customError};
+        }
+
         // Fetch the attachments of the meeting.
         gcalendar:Attachment[]|error? calendarEventAttachments = calendar:getCalendarEventAttachments(
-                meeting.googleEventId);
+                meeting.googleEventId, creatorEmailResult);
         if calendarEventAttachments is error {
             string customError = string `Error occurred while fetching the attachments!`;
             log:printError(customError, calendarEventAttachments);
@@ -745,9 +759,21 @@ service http:InterceptableService / on new http:Listener(9090) {
             };
         }
 
+        string|error? creatorEmailResult = database:getEventCreatorByGoogleEventId(meeting.googleEventId);
+        if creatorEmailResult is error {
+            string customError = "Error occurred while fetching the event creator!";
+            log:printError(customError, creatorEmailResult);
+            return <http:InternalServerError>{body: customError};
+        }
+        if creatorEmailResult is null {
+            string customError = "Event creator not found!";
+            log:printError(customError);
+            return <http:InternalServerError>{body: customError};
+        }
+
         // Delete the meeting from the calendar.
         calendar:DeleteCalendarEventResponse|error deleteCalendarEventResponse = calendar:deleteCalendarEvent(
-                meeting.googleEventId);
+                meeting.googleEventId, creatorEmailResult);
         if deleteCalendarEventResponse is error {
             string customError = string `Error occurred while deleting the meeting!`;
             log:printError(customError, deleteCalendarEventResponse);
